Consolidate embed protocol tests into one booted viewer
